Add files via upload
añadidos condiciones para el goal

+++ src/casa/suelo/OrinarSueloTerminarPlan.java
@@ -5,7 +5,6 @@ import java.util.ArrayList;
 import jadex.adapter.fipa.SFipa;
 import jadex.runtime.IMessageEvent;
 import jadex.runtime.Plan;
-import jadex.runtime.impl.RMessageEvent;
 import ontologia.acciones.OrinarSuelo;
 import ontologia.conceptos.necesidades.Diversion;
 import ontologia.conceptos.necesidades.Higiene;
@@ -17,13 +16,13 @@ public class OrinarSueloTerminarPlan extends Plan {
 
 	@SuppressWarnings("unchecked")
 	@Override
-	public void body() { 
-//		if(((ArrayList<IMessageEvent>) getBeliefbase().getBelief("mensajes_orinar_suelo").getFact())).size() == 1){
-//			getGoalbase().getGoal("terminar_orinar_suelo").drop();
-//		}
-		getGoalbase().getGoal("terminar_orinar_suelo").drop();
+	public void body() {
+		if (((ArrayList<IMessageEvent>) getBeliefbase().getBelief("mensajes_orinar_suelo").getFact()).size() == 1) {
+			getGoalbase().getGoal("terminar_orinar_suelo").drop();
+		}
 
-		IMessageEvent peticion = (IMessageEvent) ((ArrayList<IMessageEvent>)getBeliefbase().getBelief("mensajes_orinar_suelo").getFact()).get(0);
+		IMessageEvent peticion = ((ArrayList<IMessageEvent>) getBeliefbase()
+				.getBelief("mensajes_orinar_suelo").getFact()).get(0);
 		OrinarSuelo content = (OrinarSuelo) peticion.getContent();
 		HasOrinadoSuelo response = new HasOrinadoSuelo();
 
@@ -39,21 +38,21 @@ public class OrinarSueloTerminarPlan extends Plan {
 		v.setGrado(content.getVejiga().getGrado() + Necesidad.NC_NORMAL);
 		response.setVejiga(v);
 
-		ArrayList<IMessageEvent> arrayMensajes = (ArrayList<IMessageEvent>)getBeliefbase().getBelief("mensajes_orinar_suelo").getFact();
+		ArrayList<IMessageEvent> arrayMensajes = (ArrayList<IMessageEvent>) getBeliefbase()
+				.getBelief("mensajes_orinar_suelo").getFact();
 		arrayMensajes.remove(0);
 		getBeliefbase().getBelief("mensajes_orinar_suelo").setFact(arrayMensajes);
-		
-		ArrayList<Integer> arrayTiempos = (ArrayList<Integer>)getBeliefbase().getBelief("tiempos_orinar_suelo").getFact();
+
+		ArrayList<Integer> arrayTiempos = (ArrayList<Integer>) getBeliefbase().getBelief("tiempos_orinar_suelo")
+				.getFact();
 		arrayTiempos.remove(0);
 		getBeliefbase().getBelief("tiempos_orinar_suelo").setFact(arrayTiempos);
-		
+
 		IMessageEvent inform = createMessageEvent("has_orinado_suelo");
 		inform.getParameterSet(SFipa.RECEIVERS).addValue(peticion.getParameterSet(SFipa.SENDER).getValues());
 		inform.setContent(response);
 		sendMessage(inform);
 
-		
-		
 	}
 
 }
